theme rendering as regular render strategy
Use a more generic approach to rendering "to the UI" instead of generating
Components. Solo now renders the changed file and the solo file on top of it -
this allows for updated views on files where a dependency is changed (e.g.
headerfooter.gtpl).  Also disabled caching for Groovy templates.